Update to modern Rust
This reorganizes the Gaol library to comply more with modern Rust
standards, also updating it to the latest Rust edition.

The only functional change was the addition of the `open` and `openat`
syscalls to the allowed syscalls for file reading, to fix the existing
tests that attempt to open a file.